Refactor cli tool
