+++ api/opentrons/deck_calibration/dc_main.py
@@ -12,7 +12,7 @@ from numpy import dot, array
 from opentrons.robot import robot_configs
 from opentrons import robot, instruments
 from opentrons.util.calibration_functions import probe_instrument
-from opentrons.cli.linal import solve, add_z, apply_transform
+from opentrons.deck_calibration.linal import solve, add_z, apply_transform
 
 
 # Application constants
