+++ compute/scripts/calibrate
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-python -m opentrons.cli.main
+python -m opentrons.deck_calibration.dc_main
